+++ tutorials/tsom/kura_data/tsom_kura_missing_fitting.py
@@ -1,6 +1,6 @@
 import matplotlib.pyplot as plt
 from libs.datasets.artificial.kura_tsom import load_kura_tsom
-from libs.models.tsom_missing_value import TSOM2
+from libs.models.tsom import TSOM2
 from mpl_toolkits.mplot3d import Axes3D
 import matplotlib.animation as animation
